combine landing page and early access page cypress tests
